Claim all content in separate component
